% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{This function visualizes the distribution of the test-statistic of choice,
using the output of any of the 'simulate_...()' functions.}
\usage{
plot_distribution(df, ref_val = "none")
}
\description{
This function visualizes the distribution of the test-statistic of choice,
using the output of any of the 'simulate_...()' functions.
}
