% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_slope_differences.R
\name{resample_slope_differences}
\alias{resample_slope_differences}
\title{Creates a data-frame with differences between regression slopes of two
levels of a categorical variable, using permuting or bootstrapping The input
data needs to have one categorical predictor variable, one continuous predictor
variable and one continuous response variable. The function fits a linear
regression model with lm(). It is your responsibility to explore the data and
check all linear regression assumptions before using this function. Any data
transformation has to be done before using this function. Output is a dataframe
with two columns: replicate and stats. The latter  is the difference between
the estimated regression slopes for the two levels of the categorical predictor
variable.}
\usage{
resample_slope_differences(
  mydata = forest_succession,
  continuous_variable = "age",
  categorical_variable = "top_pos",
  response_variable = "sr",
  type = "permute",
  repl = 1000
)
}
\description{
Creates a data-frame with differences between regression slopes of two
levels of a categorical variable, using permuting or bootstrapping The input
data needs to have one categorical predictor variable, one continuous predictor
variable and one continuous response variable. The function fits a linear
regression model with lm(). It is your responsibility to explore the data and
check all linear regression assumptions before using this function. Any data
transformation has to be done before using this function. Output is a dataframe
with two columns: replicate and stats. The latter  is the difference between
the estimated regression slopes for the two levels of the categorical predictor
variable.
}
