% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{This function runs a simulation model multiple times and stores the statistics
of interest in a dataframe. The input data and model specifications are generated
by the function 'specify_simulation()'. In this function you define the
number pf replicates, the sample size in each simulation run, and the distribution
from which the x-values are randomly sampled.}
\usage{
run_simulation(
  input_data = mydf,
  reps = 1000,
  sample_size = "as_data",
  x_distr = "as_data"
)
}
\description{
This function runs a simulation model multiple times and stores the statistics
of interest in a dataframe. The input data and model specifications are generated
by the function 'specify_simulation()'. In this function you define the
number pf replicates, the sample size in each simulation run, and the distribution
from which the x-values are randomly sampled.
}
