% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_slope_differences.R
\name{simulate_slope_differences}
\alias{simulate_slope_differences}
\title{Creates a data-frame with differences between regression slopes of two
levels of a categorical variable, using simulation approach (see below). The
input data needs to have one categorical predictor variable, one continuous
predictor variable and one continuous response variable. The function fits a
linear regression model with lm(). It is your responsibility to explore the
data and check all linear regression assumptions before using  this function.
Any data transformation has to be done before using this function.
The simulation modeling is done as follows:
The continuous predictor values are sampled from the observed range of values,
with sample probabilities weighted using the density curve fitted on the
observed values. The two levels of the categorical predictor values are
sampled with probabilities equal to the relative frequencies of the levels
in the original data. The number of predictor values sampled is equal to the
number data points in the original data. Response variable data is obtained
by adding prediction errors to a slope of zero (type = simulate_0) or to the
predicted y-values (type = simulate_ci). Both the residuals and predicted
values are computed from the regression model that is fitted on the original
data. Output is a dataframe with two columns: replicate and stats. The latter
is the difference between the estimated regression slopes for the two levels
of the categorical predictor variable.}
\usage{
simulate_slope_differences(
  mydata = forest_succession,
  continuous_variable = "age",
  categorical_variable = "top_pos",
  response_variable = "sr",
  type = "null-hypothesis",
  repl = 1000,
  sample_size = 1,
  error_term = 1,
  equal_variance = "yes"
)
}
\description{
Creates a data-frame with differences between regression slopes of two
levels of a categorical variable, using simulation approach (see below). The
input data needs to have one categorical predictor variable, one continuous
predictor variable and one continuous response variable. The function fits a
linear regression model with lm(). It is your responsibility to explore the
data and check all linear regression assumptions before using  this function.
Any data transformation has to be done before using this function.
The simulation modeling is done as follows:
The continuous predictor values are sampled from the observed range of values,
with sample probabilities weighted using the density curve fitted on the
observed values. The two levels of the categorical predictor values are
sampled with probabilities equal to the relative frequencies of the levels
in the original data. The number of predictor values sampled is equal to the
number data points in the original data. Response variable data is obtained
by adding prediction errors to a slope of zero (type = simulate_0) or to the
predicted y-values (type = simulate_ci). Both the residuals and predicted
values are computed from the regression model that is fitted on the original
data. Output is a dataframe with two columns: replicate and stats. The latter
is the difference between the estimated regression slopes for the two levels
of the categorical predictor variable.
}
