% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_distribution.R
\name{visualise_distribution}
\alias{visualise_distribution}
\title{This function visualizes the distribution of the test-statistic of choice,
using the output of any of the 'simulate_...()' functions.}
\usage{
visualise_distribution(
  mydata = example_slope_differences,
  test_stat = "slope_difference",
  ref_val = "none"
)
}
\description{
This function visualizes the distribution of the test-statistic of choice,
using the output of any of the 'simulate_...()' functions.
}
