% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_simulation.R
\name{visualise_simulation}
\alias{visualise_simulation}
\title{This function visualizes the result of a single simulation, plotting the original
data and the simulated data in one single graph. Data and the simulation settings
are taken from the data frame + meta data that is generated by the function
'specify_simulation()'. In this function, a few choices need to be made with regard
to the sampling procedure.}
\usage{
visualise_simulation(df, xdistr = "as data", sample_size = "as data")
}
\description{
This function visualizes the result of a single simulation, plotting the original
data and the simulated data in one single graph. Data and the simulation settings
are taken from the data frame + meta data that is generated by the function
'specify_simulation()'. In this function, a few choices need to be made with regard
to the sampling procedure.
}
